file structure of state packages
